# swagger-full.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "图书馆座位预约系统 API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "开发环境"
        },
        {
            "url": "http://localhost:3001",
            "description": "测试环境"
        },
        {
            "url": "http://localhost:3002",
            "description": "UAT 环境"
        },
        {
            "url": "http://localhost:3003",
            "description": "生产环境"
        }
    ],
...
    "paths": {
        "/api/auth/wxlogin": {
            "post": {
                "tags": ["🔐 认证授权 (Auth)"],
                "summary": "微信小程序登录",
                "description": "使用微信 code 换取 JWT Token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["code"],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "【必填】微信登录 code"
                                    },
                                    "userInfo": {
                                        "type": "object",
                                        "description": "【可选】用户信息",
                                        "properties": {
                                            "avatarUrl": {
                                                "type": "string",
                                                "format": "uri"
                                            },
                                            "nickName": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "description": "JWT Token"
                                                },
                                                "userInfo": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "缺少 code": {
                                        "value": {
                                            "success": false,
                                            "error": {
                                                "code": 2001,
                                                "message": "缺少微信登录 code"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/bindStudentId": {
            "post": {
                "tags": ["🔐 认证授权 (Auth)"],
                "summary": "绑定学号",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["studentId", "realName"],
                                "properties": {
                                    "studentId": {
                                        "type": "string",
                                        "description": "【必填】学号"
                                    },
                                    "realName": {
                                        "type": "string",
                                        "description": "【必填】真实姓名"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "绑定成功"
                    },
                    "400": {
                        "description": "参数错误"
                    },
                    "409": {
                        "description": "学号已被绑定",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 2006
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "该学号已被其他用户绑定"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/check": {
            "get": {
                "tags": ["🔐 认证授权 (Auth)"],
                "summary": "检查登录状态",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已登录",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "isLogin": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "userInfo": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授权"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Token 认证，格式：Bearer <token>"
            }
        },
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "integer",
                                "example": 1001,
                                "description": "数字错误码"
                            },
                            "message": {
                                "type": "string",
                                "example": "错误描述"
                            }
                        }
                    }
                },
                "required": ["success", "error"]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "objectId",
                        "description": "用户 ID"
                    },
                    "username": {
                        "type": "string",
                        "description": "用户名"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true,
                        "description": "用户姓名"
                    },
                    "avatar": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true,
                        "description": "头像 URL"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true,
                        "description": "电子邮件"
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "description": "手机号"
                    },
                    "studentId": {
                        "type": "string",
                        "nullable": true,
                        "description": "学号"
                    },
                    "role": {
                        "type": "integer",
                        "enum": [0, 1],
                        "description": "用户角色"
                    },
                    "isSuperAdmin": {
                        "type": "boolean",
                        "description": "是否超级管理员"
                    },
                    "creditScore": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100,
                        "description": "信用积分"
                    },
                    "blacklisted": {
                        "type": "boolean",
                        "description": "是否黑名单用户"
                    },
                    "blacklistReason": {
                        "type": "string",
                        "nullable": true,
                        "description": "黑名单原因"
                    },
                    "favorites": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "收藏座位 ID 列表"
                    },
                    "activityRegistrations": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "activity": {
                                    "type": "string",
                                    "description": "活动 ID"
                                },
                                "registeredAt": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "报名时间"
                                }
                            }
                        },
                        "description": "活动报名记录"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "创建时间"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "更新时间"
                    }
                }
            },
...
        }
    }
}